requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
openai==1.3.5
supabase==1.0.3
yt-dlp==2023.3.4
whisper==1.1.10
Werkzeug==2.0.1
SQLAlchemy==1.4.23